backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
SQLAlchemy==2.0.34
alembic==1.13.2
pydantic==2.8.2
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
httpx==0.27.0
slowapi==0.1.9
email-validator==2.2.0
python-multipart==0.0.20